refactoring route /-/v1/app
